requirements.txt
Pillow>=10.0.0
google-generativeai>=0.8.0
pytest>=8.0.0
